refactor: reorganize plugin-related code by moving functions to dedicated files and updating imports
