+++ packages/snap/src/plugins/index.ts
@@ -1,0 +1,1 @@
+export { processPlugins } from './process-plugins'
