+++ packages/snap/src/plugins/plugin-dependencies.ts
@@ -1,0 +1,7 @@
+export const pluginDependencies = [
+  '@motiadev/core',
+  '@motiadev/plugin-logs',
+  '@motiadev/plugin-states',
+  '@motiadev/plugin-endpoint',
+  '@motiadev/plugin-observability',
+]
