+++ packages/snap/src/start.ts
@@ -2,7 +2,7 @@ import { createEventManager, createServer, createStateAdapter, type MotiaPlugin,
 import path from 'path'
 import { workbenchBase } from './constants'
 import { generateLockedData, getStepFiles } from './generate-locked-data'
-import { processPlugins } from './generate-plugins'
+import { processPlugins } from './plugins'
 import { activatePythonVenv } from './utils/activate-python-env'
 import { version } from './version'
